Module naming: disambiguation and improvements
