+++ sample-app-feature2/business-usecase/src/main/java/dev/marcosalis/feature2/business/usecase/Feature2UseCase.kt
@@ -1,0 +1,9 @@
+package dev.marcosalis.feature2.business.usecase
+
+import dev.marcosalis.clean.feature2.entity.Feature2Action
+
+interface Feature2UseCase {
+
+    suspend fun doStuff(payload: Feature2Action)
+
+}
