+++ sample-app-feature2/business-usecase/src/main/java/dev/marcosalis/feature2/business/usecase/di/UseCaseFeature2Module.kt
@@ -1,0 +1,19 @@
+@file:Suppress("unused")
+
+package dev.marcosalis.feature2.business.usecase.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import dev.marcosalis.feature2.business.usecase.Feature2UseCase
+import dev.marcosalis.feature2.business.usecase.Feature2UseCaseImpl
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal interface UseCaseFeature2Module {
+
+    @Binds
+    fun feature1(impl: Feature2UseCaseImpl): Feature2UseCase
+
+}
