--- sample-app-feature2/business/build.gradle.kts
@@ -1,55 +1,0 @@
-
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
-
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.dagger.hilt.android)
-    alias(libs.plugins.ksp)
-}
-
-android {
-    namespace = "dev.marcosalis.feature2.business"
-
-    compileSdk {
-        version = release(libs.versions.sdk.compile.get().toInt())
-    }
-
-    defaultConfig {
-        minSdk = libs.versions.sdk.min.get().toInt()
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_21
-        targetCompatibility = JavaVersion.VERSION_21
-    }
-}
-
-tasks.withType<KotlinJvmCompile>().configureEach {
-    compilerOptions {
-        jvmTarget.set(JvmTarget.JVM_21)
-    }
-}
-
-dependencies {
-    api(project(":sample-app-feature2:entity"))
-    implementation(project(":sample-app-feature2:data-access"))
-
-    implementation(libs.androidx.core.ktx)
-
-    // Dagger / Hilt
-    implementation(libs.hilt.android)
-    ksp(libs.hilt.compiler)
-
-    implementation(libs.timber)
-
-    testImplementation(libs.junit)
-}
