--- sample-app-feature2/business/src/main/java/dev/marcosalis/feature2/business/Feature2UseCase.kt
@@ -1,9 +1,0 @@
-package dev.marcosalis.feature2.business
-
-import dev.marcosalis.clean.feature2.entity.Feature2Action
-
-interface Feature2UseCase {
-
-    suspend fun doStuff(payload: Feature2Action)
-
-}
