+++ sample-business-usecase/src/main/kotlin/dev/marcosalis/clean/business/usecase/init/InitializeAppUseCase.kt
@@ -1,0 +1,7 @@
+package dev.marcosalis.clean.business.usecase.init
+
+interface InitializeAppUseCase {
+
+    suspend fun initialize()
+
+}
