+++ sample-business-usecase/src/test/kotlin/dev/marcosalis/clean/business/usecase/ExampleUnitTest.kt
@@ -1,0 +1,16 @@
+package dev.marcosalis.clean.business.usecase
+
+import org.junit.Assert
+import org.junit.Test
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        Assert.assertEquals(4, 2 + 2)
+    }
+}
