--- sample-business/src/main/kotlin/dev/marcosalis/clean/business/feature1/Feature1UseCase.kt
@@ -1,13 +1,0 @@
-package dev.marcosalis.clean.business.feature1
-
-import dev.marcosalis.clean.business.entity.feature1.Feature1
-import dev.marcosalis.clean.business.entity.feature1.Feature1Action
-import kotlinx.coroutines.flow.Flow
-
-interface Feature1UseCase {
-
-    val data: Flow<Feature1>
-
-    suspend fun performAction(payload: Feature1Action)
-
-}
