--- sample-business/src/main/kotlin/dev/marcosalis/clean/business/init/InitializeAppUseCase.kt
@@ -1,7 +1,0 @@
-package dev.marcosalis.clean.business.init
-
-interface InitializeAppUseCase {
-
-    suspend fun initialize()
-
-}
