--- sample-business/src/test/kotlin/dev/marcosalis/clean/business/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package dev.marcosalis.clean.business
-
-import org.junit.Assert
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        Assert.assertEquals(4, 2 + 2)
-    }
-}
