+++ settings.gradle.kts
@@ -23,14 +23,14 @@ dependencyResolutionManagement {
 
 rootProject.name = "Android Clean Architecture"
 include(":sample-app")
-include(":sample-business")
+include(":sample-business-usecase")
 include(":sample-data-access")
 include(":sample-data")
 include(":sample-data:injection")
 include(":sample-entity")
 include(":sample-ktx")
 include(":sample-app-feature2")
-include(":sample-app-feature2:business")
+include(":sample-app-feature2:business-usecase")
 include(":sample-app-feature2:data-access")
 include(":sample-app-feature2:data")
 include(":sample-app-feature2:entity")
